+++ CHANGELOG.md
@@ -11,6 +11,7 @@
  - feat: add `NesIcons.screen`.
  - feat: add `NesIcons.rename`.
  - feat: add `NesIcons.hamburgerMenu`.
+ - feat: adding `NesButton` helper constructors.
 
 # 0.12.1
  - fix: theme lerp causing error on null access.