+++ CHANGELOG.md
@@ -13,6 +13,7 @@
  - feat: add `NesIcons.hamburgerMenu`.
  - feat: adding `NesButton` helper constructors.
  - feat: add `NesIcons.arrowCursor`.
+ - feat: `NesWindow` uses `NesContainerTheme` instead of Material's card.
 
 # 0.12.1
  - fix: theme lerp causing error on null access.