+++ CHANGELOG.md
@@ -15,6 +15,7 @@
  - feat: add `NesIcons.arrowCursor`.
  - feat: `NesWindow` uses `NesContainerTheme` instead of Material's card.
  - feat: update mini sprite to improve icons.
+ - feat: add `NesIcons.center`.
 
 # 0.12.1
  - fix: theme lerp causing error on null access.